+++ src/app/_models/variable.ts
@@ -2,5 +2,5 @@
     id: number;
     name: string;
     value: string;
-    dashboard_id: number;
+    container_id: number;
 }